requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
